Overhaul simulation technique & remove buggy code
After looking back over the old code for simulating coalescing ballistic
annihilation in the tradtional way and sampling geometric random variables,
it seems there are still bugs.

The most prominent is that under the previous method of simulating, each
"step", all of the arrow-arrow collisions (i.e. > <) collisions were resolved
before any potential arrow-blockade-arrow collisions (i.e. > * * * ... * * <).

The issue with this approach is that it can sometimes be incorrect to always do
arrow-arrow collisions first, which I didn't realize before. Example:

>     *   *   *   >           <
|--2--|-1-|-1-|-1-|-----5-----|

In the above situation, my algorithm would first resolve the arrow-arrow collision
on the far right. And supposing the left moving arrow survies the collision:

>     *   *   *   <
|--2--|-1-|-1-|-1-|

which would be wrong, as the arrow on the far left would reach the
blockade in front of it before the arrow-arrow collision on the far right
would occur. But in my mistaken algorithm, the arrow on the far right would
get to hit the blockade in front of it FIRST, before the arrow on the far left
got to hit the blockade in front of it, which is incorrect.

The approach of this new algorithm is slower but more correct, as we always resolve
the closest collision no matter where it occurs on our "infinite" number line. Thus
situations like the above example and more are simulated correctly.

Future improvements would be to develop a faster algorithm, but this algorithm
will be used going forward as I value correctness over speed. And it still simulates
~100K particles in a reasonable time, which should be enough to be significant.